Module parity_crypto::publickey::ec_math_utils

Multiple primitives for work with public and secret keys and with secp256k1 curve points

Structs

CURVE_ORDER

Constants

BASE_POINT_BYTES

Generation point array combined from X and Y coordinates Equivalent to uncompressed form, see https://tools.ietf.org/id/draft-jivsov-ecc-compact-05.html#rfc.section.3

Functions

generation_point

Return the generation point (aka base point) of secp256k1

public_add

In-place add one public key to another (EC point + EC point)

public_mul_secret

In-place multiply public key by secret key (EC point * scalar)

public_negate

Replace a public key with its additive inverse (EC point = - EC point)

public_sub

In-place sub one public key from another (EC point - EC point)